# mcp-server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MCP Kafka Server API",
    "version": "1.0.0",
    "description": "API for interacting with a Strimzi Kafka cluster via an MCP server."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Returns the operational status of the server.",
        "responses": {
          "200": {
            "description": "Server is operational",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "UP"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/topics": {
      "get": {
        "summary": "List Kafka Topics",
        "description": "Retrieves a list of all topic names from the connected Kafka cluster.",
        "responses": {
          "200": {
            "description": "A list of topic names.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "topics": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "topic1",
                        "my-new-topic"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error (e.g., failed to connect to Kafka)"
          }
        }
      },
      "post": {
        "summary": "Create Kafka Topic (via Strimzi)",
        "description": "Creates a new Kafka topic by creating a KafkaTopic custom resource in Kubernetes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "topicName": {
                    "type": "string",
                    "description": "The desired name for the new topic.",
                    "example": "my-new-topic-from-api"
                  },
                  "partitions": {
                    "type": "integer",
                    "description": "Number of partitions for the topic (default: 1).",
                    "example": 3
                  },
                  "replicas": {
                    "type": "integer",
                    "description": "Replication factor for the topic (default: 1).",
                    "example": 1
                  }
                },
                "required": [
                  "topicName"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "KafkaTopic resource created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "kind": "KafkaTopic",
                    "metadata": {
                      "name": "my-new-topic-from-api"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., missing topicName)."
          },
          "409": {
            "description": "Conflict (Topic resource already exists)."
          },
          "500": {
            "description": "Internal server error (e.g., failed to interact with Kubernetes API)."
          }
        }
      }
    },
    "/api/produce": {
      "post": {
        "summary": "Produce Messages to Topic",
        "description": "Sends one or more messages to the specified Kafka topic.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "topic": {
                    "type": "string",
                    "description": "The target topic name.",
                    "example": "my-new-topic"
                  },
                  "messages": {
                    "type": "array",
                    "description": "An array of messages to produce.",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string",
                          "description": "Optional message key.",
                          "example": "messageKey123"
                        },
                        "value": {
                          "type": "string",
                          "description": "The message content.",
                          "example": "Hello Kafka!"
                        }
                      },
                      "required": [
                        "value"
                      ]
                    }
                  }
                },
                "required": [
                  "topic",
                  "messages"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Messages produced successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Messages produced successfully"
                    },
                    "recordMetadata": {
                      "type": "array",
                      "example": [
                        {
                          "topicName": "my-new-topic",
                          "partition": 0,
                          "errorCode": 0,
                          "offset": "10"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., missing topic or messages)."
          },
          "500": {
            "description": "Internal server error (e.g., failed to produce to Kafka)."
          }
        }
      }
    },
    "/api/consumer-groups": {
      "get": {
        "summary": "Describe Consumer Groups",
        "description": "Retrieves details about specified consumer groups from Kafka.",
        "parameters": [
          {
            "name": "groupIds",
            "in": "query",
            "required": true,
            "description": "Comma-separated list of consumer group IDs to describe.",
            "schema": {
              "type": "string"
            },
            "example": "my-consumer-group,another-group"
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the specified consumer groups.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "groups": [
                      {
                        "groupId": "my-consumer-group",
                        "state": "Stable"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., missing groupIds query parameter)."
          },
          "500": {
            "description": "Internal server error (e.g., failed to describe groups in Kafka)."
          }
        }
      }
    },
    "/api/groups": {
      "get": {
        "summary": "List Consumer Groups",
        "description": "Retrieves a list of all consumer group IDs from the Kafka cluster.",
        "operationId": "listConsumerGroups",
        "tags": ["Consumer Groups"],
        "responses": {
          "200": {
            "description": "A list of consumer group IDs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": ["my-group-1", "another-group"]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/groups/{groupId}/lag": {
      "get": {
        "summary": "Get Consumer Group Lag",
        "description": "Calculates the lag for a specific consumer group. Lag is the difference between the latest message offset in a partition and the offset committed by the consumer group.",
        "operationId": "getConsumerGroupLag",
        "tags": ["Consumer Groups"],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "description": "The ID of the consumer group.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topics",
            "in": "query",
            "required": false,
            "description": "Optional comma-separated list of topic names to filter the lag calculation.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed lag information for the consumer group.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "groupId": {
                      "type": "string"
                    },
                    "totalLag": {
                      "type": "integer",
                      "format": "int64"
                    },
                    "topics": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "object",
                          "properties": {
                            "lag": {
                              "type": "integer",
                              "format": "int64",
                              "description": "Calculated lag for the partition (-1 if unknown)."
                            },
                            "consumerOffset": {
                              "type": "string",
                              "description": "The last offset committed by the consumer group."
                            },
                            "latestOffset": {
                              "type": "string",
                              "description": "The latest offset (high water mark) in the partition."
                            }
                          }
                        }
                      }
                    }
                  },
                  "example": {
                    "groupId": "my-consumer-group",
                    "totalLag": 150,
                    "topics": {
                      "my-topic": {
                        "0": {
                          "lag": 100,
                          "consumerOffset": "500",
                          "latestOffset": "600"
                        },
                        "1": {
                          "lag": 50,
                          "consumerOffset": "250",
                          "latestOffset": "300"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Consumer group not found or not active."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "List Kafka Users (Strimzi)",
        "description": "Retrieves a list of KafkaUser custom resources from the Kubernetes cluster.",
        "operationId": "listKafkaUsers",
        "tags": ["Kafka Users"],
        "responses": {
          "200": {
            "description": "A list of Kafka users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                       "name": { "type": "string" },
                       "status": { "type": "object", "description": "Status subresource if available" }
                    }
                  },
                  "example": [
                    { "name": "my-user", "status": { "observedGeneration": 1, "conditions": [{"type": "Ready", "status": "True"}] } },
                    { "name": "another-app-user", "status": { "observedGeneration": 1, "conditions": [{"type": "Ready", "status": "True"}] } }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (e.g., failed to connect to Kubernetes API)"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Topics",
      "description": "Operations related to Kafka topics"
    },
    {
      "name": "Producer",
      "description": "Send messages to Kafka topics"
    },
    {
      "name": "Consumer Groups",
      "description": "Operations related to Kafka consumer groups"
    },
    {
      "name": "Kafka Users",
      "description": "Operations related to Strimzi KafkaUser custom resources"
    }
  ]
}